delete internal scheduler
