+++ core/shared/src/main/scala/zio/scheduler/package.scala
@@ -16,8 +16,6 @@
 
 package zio
 
-import zio.internal.{ Scheduler => IScheduler }
-
 package object scheduler {
   type Scheduler = Has[Scheduler.Service]
